break down email thread into smaller components
